Fixed errors and started ticket 9
