+++ src/App.js
@@ -4,10 +4,10 @@ import Header from "./components/header.jsx";
 import Reviews from "./components/reviews";
 import ReviewCard from "./components/reviewCard";
 import { Link } from "react-router-dom";
-import Comments from "./components/comments";
+import Comments from "./components/commentsSection";
 import SignInPage from "./components/signInPage";
 import { useState } from "react";
-import 'bootstrap/dist/css/bootstrap.min.css';
+import "bootstrap/dist/css/bootstrap.min.css";
 
 function App() {
     const [user, setUser] = useState({});
@@ -23,9 +23,9 @@ function App() {
                     element={<SignInPage user={user} setUser={setUser} signedIn={signedIn} setSignedIn={setSignedIn} />}
                 />
                 <Route path="/" element={<Reviews />} />
-                <Route path="/reviews" element={<Reviews signedIn={signedIn} user={user}/>} />
+                <Route path="/reviews" element={<Reviews signedIn={signedIn} user={user} />} />
                 <Route path="/reviews/:review_id" element={<ReviewCard user={user} signedIn={signedIn} />} />
-                <Route path="/reviews/:review_id" element={<Comments user={user} signedIn={signedIn}/>} />
+                <Route path="/reviews/:review_id" element={<Comments user={user} signedIn={signedIn} />} />
             </Routes>
         </div>
     );
